A0-2264: Plan error boundaries around components dealing with unreliable data
